third_party/WebKit/LayoutTests/http/tests/security/link-crossorigin-subresource-use-credentials-expected.txt

CONSOLE ERROR: Link subresource from origin 'http://localhost:8080' has been blocked from loading by Cross-Origin Resource Sharing policy: A wildcard '*' cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Link subresource from origin 'http://localhost:8080' has been blocked from loading by Cross-Origin Resource Sharing policy: A wildcard '*' cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test that a subresource with a crossorigin=use-credentials attribute loads the expected CORS enabled resources.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 
PASS 
PASS 
PASS 
PASS successfullyParsed is true

TEST COMPLETE
